+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+- Added [BlocWithEffectsObserver] with [onEffect] callback
+- [useEffect] throws [StateError] if state is closed
+ 
 ## 0.1.1
 - Updated documentation, example, readme [README.md]